pyproject.toml
[tool.poetry]
name = "wb-feedbacks"
version = "0.1.0"
description = ""
authors = ["Yuhen Liubimov <yauhenliubimau@gmail.com>"]
readme = "README.md"
packages = [{include = "wb_feedbacks"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
openpyxl = "^3.1.2"
schedule = "^1.2.1"
environs = "^11.0.0"
python-telegram-bot = "^21.0.1"
types-requests = "^2.31.0.20240311"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"
mypy = "^1.9.0"

[tool.ruff]
cache-dir = ".venv/ruff/"
exclude = [
    ".git",
    ".git-rewrite",
    ".pyenv",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "site-packages",
]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.mypy]
cache_dir = ".venv/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = true
exclude = [
    ".venv",
    ".github",
    ".idea",
    ".vscode",
    "venv",
]
files = ["."]
pretty = true
python_version = "3.10"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
